Respond to review requests.
